Updated Controllers for item logic and login, logout methods

+++ wishlist_dotnetcore/WishList/Controllers/ItemController.cs
@@ -1,21 +1,29 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var currentUser = _userManager.GetUserAsync(User).Result;
+
+            var model = _context.Items.Where(item => item.User.Id == currentUser.Id).ToList();
 
             return View("Index", model);
         }
@@ -29,6 +37,8 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            var currentUser = _userManager.GetUserAsync(User).Result;
+            item.User = currentUser;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -36,7 +46,18 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+            var currentUser = _userManager.GetUserAsync(User).Result;
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if(item==null)
+            {
+                return NotFound();
+            }
+
+            if(item.User.Id!=currentUser.Id)
+            {
+                return Unauthorized();
+            }
+
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
